Suppress false positive from static analysis
Static analysis flags this line as a syntax error, but this is probably
due to the complexity of the macros used. The exact same line is used
in other places in the same file, and is not flagged as an error.

This commit adds a suppress command to tel cppcheck to ignore this
error.